feat: Выполнить ТЗ 15 спринта v00 - добавить request - добавить тесты

+++ src/main/java/ru/practicum/shareit/user/controller/UserController.java
@@ -30,7 +30,7 @@ public class UserController {
     }
 
     @PatchMapping("/{userId}")
-    public User updateUser(@RequestBody UserDto userDto, @PathVariable int userId) {
+    public User updateUser(@RequestBody UserDto userDto, @PathVariable long userId) {
         return userService.updateUser(userDto, userId);
     }
